:zap: 업다운 게임 실행할 때 렌더링 최적화
