+++ client/src/components/room/games/index.tsx
@@ -1,35 +1,27 @@
-import React, { useMemo } from 'react';
-import { Wrapper } from '@components/room/games/index.style';
-import GameBox from '@components/room/games/GameBox';
-import { GameIcon } from '@components/icons';
+import React, { useEffect } from 'react';
+import { useSelector } from 'react-redux';
+import { RootState } from '@src/store';
+import UpdownGame from '@components/room/games/UpdownGame';
+import useGameSocket from '@hooks/socket/useGameSocket';
 
 type GamesPropTypes = {
-  GameStartHandlerList: [];
+  startGamesRef: React.MutableRefObject<Object>;
 };
-const Games: React.FC<GamesPropTypes> = ({ GameStartHandlerList }) => {
-  const GameList = useMemo(
-    () => [
-      //아이콘, 설명은 임시
-      { icon: <GameIcon />, title: '랜덤뽑기', explain: '설명' },
-      { icon: <GameIcon />, title: '업다운', explain: '설명' },
-      { icon: <GameIcon />, title: '라이어게임', explain: '설명' },
-    ],
-    [],
-  );
 
-  return (
-    <Wrapper>
-      {GameList.map(({ icon, title, explain }, index) => (
-        <GameBox
-          key={index}
-          icon={icon}
-          title={title}
-          explain={explain}
-          start={GameStartHandlerList[title]}
-        />
-      ))}
-    </Wrapper>
-  );
+const Games: React.FC<GamesPropTypes> = ({ startGamesRef }) => {
+  const currentGame = useSelector((state: RootState) => state.room.currentGame.title);
+
+  const { GameStartHandlerList } = useGameSocket();
+  useEffect(() => {
+    startGamesRef.current = GameStartHandlerList;
+  }, []);
+
+  switch (currentGame) {
+    case '업다운':
+      return <UpdownGame />;
+    default:
+      return <></>;
+  }
 };
 
 export default Games;
